pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "offline-lab"
version = "0.1.0"
description = "SQL 분석 연습 및 자동 채점 플랫폼"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.9.0,<1.0.0",
    "pandas>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "streamlit>=1.30.0,<2.0.0",
    "plotly>=5.0.0,<6.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "google-genai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.hatch.build.targets.wheel]
packages = ["engine", "grader", "generator", "problems", "services", "common", "config", "dashboard", "scripts"]
